+++ plugins/postcss-unset-value/src/index.ts
@@ -1,0 +1,44 @@
+import type { PluginCreator } from 'postcss';
+import { inherited, nonInherited } from './property-def';
+
+type pluginOptions = {
+	// Preserve the original declaration.
+	// Default: false
+	preserve: boolean;
+}
+
+// Convert "unset" to "inherit" or "initial" based on the property definition.
+const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
+	const options = Object.assign({ preserve: false }, opts);
+
+	return {
+		postcssPlugin: 'postcss-unset-value',
+		Declaration(decl) {
+			if (decl.value !== 'unset') {
+				return;
+			}
+
+			let replacement : string|false = false;
+			if (inherited.has(decl.prop)) {
+				replacement = 'inherit';
+			} else if (nonInherited.has(decl.prop)) {
+				replacement = 'initial';
+			}
+
+			if (!replacement) {
+				return;
+			}
+
+			if (options.preserve) {
+				decl.cloneBefore({ prop: decl.prop, value: replacement });
+			} else {
+				decl.value = replacement;
+			}
+		},
+	};
+};
+
+creator.postcss = true;
+
+export default creator;
+