+++ sites/postcss-preset-env/postcss.config.cjs
@@ -1,0 +1,20 @@
+module.exports = ctx => {
+	const isProd = ctx.env === 'production';
+
+	return {
+		map: !isProd,
+		plugins: {
+			'postcss-import': {},
+			'postcss-preset-env': {
+				stage: 0,
+				preserve: true,
+				features: {
+					'custom-properties': false
+				}
+			},
+			'cssnano': isProd ? {
+				preset: 'default',
+			} : false,
+		},
+	};
+};