+++ plugins/postcss-color-function/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changes to PostCSS Color Function
 
-### Unreleased (patch)
+### 1.0.3 (March 8, 2022)
+
+- Fix gamut mapping giving overly unsaturated colors.
+- Implement powerless color components in gamut mapping.
+
+### 1.0.2 (February 12, 2022)
+
+- Updated `@csstools/postcss-progressive-custom-properties` to `1.1.0`.
+
+### 1.0.1 (February 11, 2022)
 
 - Add tests for percentage values in non-xyz color spaces.
 - Ignore percentage values in xyz color space as these are not supported.